crew-service/requirements.txt
crewai>=0.60.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.0.0
